refactor: 게임 진행 관련 메소드 App 클래스에서 BaseballGame 클래스로 전부 이동
게임 진행은 BaseballGame에서 관리하는 것이 더 적절하다고 판단해서 app에 작성한 로직 이동. 이동 후 연결을 수정하지 않아서 오류가 날 수 있으나 바로 수정할 것이다.
